requirements.txt
streamlit
python-dotenv
PyPDF2
faiss-cpu

langchain
langchain-community
langchain-text-splitters
langchain-google-genai

google-genai

-e .